# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book API",
        "description": "Simple CRUD API for managing books",
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "publication_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Book"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "publication_year": {
                                        "type": "integer",
                                        "example": 2020
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book by ID",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book found"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update an existing book",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Updated Author"
                                    },
                                    "publication_year": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book by ID",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreBookRequest": {
                "required": [
                    "title",
                    "author",
                    "publication_year"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "100M Leads"
                    },
                    "author": {
                        "type": "string",
                        "example": "Alex Hormozi"
                    },
                    "publication_year": {
                        "type": "integer",
                        "example": 2005
                    }
                },
                "type": "object"
            },
            "UpdateBookRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "New Book"
                    },
                    "author": {
                        "type": "string",
                        "example": "Jordan Belfort"
                    },
                    "publication_year": {
                        "type": "integer",
                        "example": 2021
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}